+++ CHANGES.md
@@ -6,6 +6,8 @@
 * Adds `bijection-protobuf`
 * Adds `TEnumCodec` for conversion of `TEnum` -> `Int`
 * Change `biject` method in `Bijection` object to `asMethod` for clarity
+* Adds default `unwrap` bijections to value case classes
+* Adds `Bijection.getOrElse` for conversions between `Option[A]` and `A`
 
 ### 0.1.0